More

+++ src/random.c
@@ -1,12 +1,14 @@
 #include "random.h"
 
-#include "esp_system.h"
+#include <esp_system.h>
+#include <stddef.h>
+#include <stdint.h>
 
 void
 random_bytes_fill (uint8_t *buffer, size_t size) {
   uint32_t value = 0;
 
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     if (i % 4 == 0) {
       value = esp_random();
     }
